airflow/requirements.txt
pyalex
pandas
numpy
lingua-language-detector
oauth2client
gspread
scikit-learn